# purification-rescue-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Purification Rescue Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analyze": {
      "post": {
        "summary": "Start Analysis",
        "description": "Starts the agent in the background and returns a Job ID immediately.",
        "operationId": "start_analysis_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Start Analysis Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/{job_id}": {
      "get": {
        "summary": "Get Status",
        "description": "Polled by the frontend to check progress.\n\n`state` is the job's lifecycle state and `progress` its latest message;\n`history` replays every message so far. `error` is set once the job has\nfailed or been interrupted. Another user's job is a 404.",
        "operationId": "get_status_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/result/{job_id}": {
      "get": {
        "summary": "Get Result",
        "description": "Retrieves the final detailed report.\n\nA stored result that no longer fits ProtocolResult (its shape changed after\nthe job ran) is returned as stored; the frontend decides whether it can\nrender it. Another user's job is a 404.",
        "operationId": "get_result_result__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trace/{job_id}": {
      "get": {
        "summary": "Get Trace",
        "description": "The job's trace (#77): each LLM call's reasoning summary, output, tokens,\ntime and estimated cost, and each stage's time. Kept out of the report so\n/result stays small. null for a job with no trace: one that ran before\ntraces existed, or is still running. Another user's job is a 404.",
        "operationId": "get_trace_trace__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Trace"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Trace Trace  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{job_id}": {
      "patch": {
        "summary": "Rename Job",
        "description": "Rename one of the user's jobs, running or not. A blank name clears it, so\nMy Jobs shows the input summary again. Another user's job is a 404.",
        "operationId": "rename_job_jobs__job_id__patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Job",
        "description": "Hard-delete one of the user's finished, failed or interrupted jobs. Another\nuser's job is a 404; a queued or running one is a 409, since jobs can't be\ncancelled (#49).",
        "operationId": "delete_job_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/delete": {
      "post": {
        "summary": "Delete Jobs",
        "description": "Delete several of the user's jobs. Running jobs and ones that aren't the\nuser's are skipped and listed with the reason, rather than failing the rest.",
        "operationId": "delete_jobs_jobs_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProtocolResult": {
        "properties": {
          "purifications": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SourceProtocol"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purifications"
          },
          "comprehensive_protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comprehensive Protocol"
          },
          "raw_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Plan"
          },
          "blast_results": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Hit"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blast Results"
          },
          "papers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Paper"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Papers"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "synthesis_skipped": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesis Skipped"
          }
        },
        "type": "object",
        "title": "ProtocolResult",
        "description": "A job's report, as stored and as /result serves it. Stored reports are never\nmigrated, so fields added later are optional; see \"Changing the\nreport\" in docs/system_runbooks/backend.md."
      },
      "PurificationRequest": {
        "properties": {
          "fasta_id": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Fasta Id",
            "description": "Protein FASTA sequence or ID"
          },
          "failed_purification_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Purification Text",
            "description": "User-provided failed purification text to override CTTdb lookup"
          },
          "min_percent_identity": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Percent Identity",
            "default": 90.0
          },
          "min_query_coverage": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Query Coverage",
            "default": 90.0
          },
          "max_evalue": {
            "type": "number",
            "maximum": 1000.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Evalue",
            "default": 1e-05
          },
          "max_hits": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Hits",
            "default": 10
          },
          "max_protocols": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Protocols",
            "description": "Source protocols to extract before the literature search stops",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "fasta_id"
        ],
        "title": "PurificationRequest"
      },
      "RenameJobRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "RenameJobRequest"
      },
...
      "Trace": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageTime"
            },
            "type": "array",
            "title": "Stages",
            "default": []
          },
          "total_seconds": {
            "type": "number",
            "title": "Total Seconds",
            "readOnly": true
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "readOnly": true
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "readOnly": true
          },
          "reasoning_tokens": {
            "type": "integer",
            "title": "Reasoning Tokens",
            "readOnly": true
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd",
            "description": "The estimated total, or None when any step's model has no price.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "total_seconds",
          "input_tokens",
          "output_tokens",
          "reasoning_tokens",
          "cost_usd"
        ],
        "title": "Trace"
      },
...
    }
  }
}